# api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Doc task todo API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/signup": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Signup",
                "operationId": "1aaf9b65f8ff26b2cd0c6c0c6c4f3803",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "username": {
                                        "type": "string",
                                        "maxLength": 32,
                                        "minLength": 6,
                                        "example": "user1"
                                    },
                                    "password": {
                                        "description": "Одна большая и маленькая латинская буква, хотя бы 1 цифра",
                                        "type": "string",
                                        "maxLength": 255,
                                        "minLength": 8,
                                        "example": "S2dJ42Msa34"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "example": "S2dJ42Msa34"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "HTTP_CREATED",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string",
                                            "example": "4|0uzyn2YlL9DbbnoC2VIMkvtLzx4jXwugXIQqPPx6927408e4"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "$ref": "#/components/responses/422"
                    }
                }
            }
        },
        "/api/signin": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Signin",
                "operationId": "35a727adb44f3109104abbb8a2d62614",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "username": {
                                        "type": "string",
                                        "maxLength": 32,
                                        "minLength": 6,
                                        "example": "user1"
                                    },
                                    "password": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "minLength": 8,
                                        "example": "S2dJ42Msa34"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "HTTP_OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string",
                                            "example": "4|0uzyn2YlL9DbbnoC2VIMkvtLzx4jXwugXIQqPPx6927408e4"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "$ref": "#/components/responses/422"
                    }
                }
            }
        },
        "/api/logout": {
            "delete": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logout",
                "operationId": "10b55471fce114a4942287b11303731f",
                "responses": {
                    "204": {
                        "description": "HTTP_NO_CONTENT"
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "responses": {
            "401": {
                "description": "HTTP_UNAUTHORIZED",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "Unauthenticated."
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "422": {
                "description": "HTTP_UNPROCESSABLE_ENTITY",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "The username has already been taken. (and 4 more errors)."
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your bearer token in the format **Bearer <token>**",
                "scheme": "bearer"
            }
        }
    }
}